pyproject.toml
[project]
name = "automaton-auditor-swarm"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "docling>=2.74.0",
    "gitpython>=3.1.46",
    "langchain>=1.2.10",
    "langchain-google-genai>=4.2.1",
    "langchain-openai>=1.1.10",
    "langgraph>=1.0.9",
    "langsmith>=0.7.6",
    "pydantic>=2.12.5",
    "python-dotenv>=1.2.1",
]
